refactor: 프로덕션 코드 패키지 구조와 일치하도록 테스트 파일 이동
