+++ src/test/java/racingcar/domain/PositionTest.java
@@ -1,8 +1,7 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Position;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
